ml1_deploy/requirements.txt
streamlit
scikit-learn
pandas
tensorflow-cpu
numpy